Clean up error handling code
