+++ packages/api/src/lib/middleware/errorHandling.ts
@@ -1,0 +1,5 @@
+export { APIErrorCodes, APIException } from './errorHandling/lib';
+export {
+  CaughtErrorHandler,
+  APIErrorHandler,
+} from './errorHandling/errorHandlers';
