+++ packages/api/src/lib/middleware/errorHandling/lib.ts
@@ -1,0 +1,27 @@
+export enum APIErrorCodes {
+  // Prisma / DB Custom errors (1xxx)
+  USER_NOT_FOUND = 1001,
+
+  // Prisma / DB Generic errors (2xxx)
+  UNIQUE_FIELD_NOT_UNIQUE = 2001,
+  ENTITY_NOT_FOUND = 2002,
+
+  // Javascript Runtime errors (3xxx)
+  JSON_FORMAT_ERROR = 3001,
+
+  // Unknown / Critical errors (9xxx)
+  UNKNOWN_ERROR = 9001,
+  PRISMA_UNKNOWN_ERROR = 9002,
+}
+
+export class APIException extends Error {
+  msg: string;
+  errorCode: APIErrorCodes;
+
+  constructor(msg: string, errorCode: APIErrorCodes) {
+    super(msg);
+
+    this.msg = msg;
+    this.errorCode = errorCode;
+  }
+}
